backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2
nltk==3.8.1
textblob==0.17.1
requests==2.31.0

# Blockchain
web3==6.11.3
eth-account==0.9.0
cryptography==41.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring
prometheus-client==0.19.0

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2
